fix: resolve all clippy warnings
- modal-common: const thread_local, regex outside loops, RangeInclusive::contains
- modal-datastore: remove unnecessary derefs, add allow for too_many_arguments
- modality-lang: Display for Path, strip_prefix, or_default, saturating_sub, matches!
- modal-node: remove unused mut, fix unwrap after is_some
- modal/modality: crate-level allows for complex patterns

All crates now pass: cargo clippy --all -- -D warnings

+++ rust/modal/src/cmds/node/ping.rs
@@ -77,7 +77,7 @@ pub async fn run(opts: &Opts) -> Result<()> {
     log::info!("--- Ping Statistics ---");
     log::info!("Total pings: {}", times_to_ping);
     log::info!("Total time: {:?}", duration);
-    log::info!("Average time: {:?}", duration / times_to_ping as u32);
+    log::info!("Average time: {:?}", duration / times_to_ping);
 
     Ok(())
 }
